rename coltypes to columntypes
